scripts/requirements.txt
databricks-sdk>=0.72.0
certifi
